Now possible to clear backend conversation
Changed the request to create/update the backend conversation from a POST to a PUT, as it fits better with what we are trying to accomplish.
Made it possible to clear the backend by the press of a button. ("Ny chat").
Added a max number of iterations for the check_for_messages function so that it wont keep going forever if something goes wrong when updating the conversation.