:bug: Fixed bug and improved backend structure
(Source/Explanation still needs to be fixed for all other than internet.)

Fixed the chat type bug.
Fix so that the frontend once again can present the messages and sources with the new backend format.
Simplify the function that checks whether the backend has updated the conversation. In the frontend we wait for get_response() to finish, as such we do not need to check more than once for the conversation. BUT, in case, it will continue checking.