:sparkles: Settings (Only for internet atm)
The internet assistant are getting and using the settings.
Choosing tools works.
Setting language level changes the system prompt, but does not seem to have an impact on the response we get.

The settings are handled in Settings.svelte.
They are stored in context in the backend.

Updated pydocs in assistant and intranet chatbots.