:arrow_right: :arrow_left: Update Message, server and assistant chatbot
The assistant chatbot have been updated to take the whole conversation instead of only the query.
The conversation in server.py now acts as our memory.
Message has been updated to add chat_type for that message as well as the settings. The user variable has been changed to role as well, to match the memory format of the chatbots.